Refactor to move GPU related classes out of the image module. get_data and set_data now raise exceptions.
