+++ examples/sprite/depth_sprite.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from pyglet.graphics import Group, ShaderProgram
-    from pyglet.image.base import TextureBase
+    from pyglet.graphics.texture import TextureBase
 
 # Standard projection Z is 0 to 255. Keep window within that.
 # You will have to change window projection if you wish to go beyond this.
