+++ pyglet/graphics/__init__.py
@@ -13,7 +13,8 @@ from pyglet.graphics.base import GeometryMode  # noqa: F401
 from pyglet.graphics import api  # noqa: F401
 from pyglet.graphics.draw import Group, ShaderGroup  # noqa: F401
 from pyglet.graphics.api import Batch, get_default_shader, get_default_batch, global_backend, Shader, ShaderProgram, ComputeShaderProgram  # noqa: F401
-
+from pyglet.graphics.texture import Texture, TextureGrid, Texture3D   # noqa: F401
+from pyglet.graphics.atlas import TextureBin, TextureArrayBin, TextureAtlas   # noqa: F401
 #
 # def _load_backend_base(name: str):
 #     if pyglet.options.backend == "opengl":
