+++ pyglet/graphics/api/gl/state.py
@@ -10,7 +10,7 @@ from pyglet.graphics.state import State
 
 if TYPE_CHECKING:
     from pyglet.graphics import Group
-    from pyglet.image.base import TextureBase
+    from pyglet.graphics.texture import TextureBase
     from pyglet.graphics.api.gl.shader import ShaderProgram
