+++ pyglet/image/__init__.py
@@ -97,39 +97,14 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, BinaryIO, Sequence
 
-import pyglet
-
-from .base import TextureFilter
-
-
 if TYPE_CHECKING:
     from pyglet.image.base import AbstractImage
 
-from pyglet.image import atlas, animation
+from pyglet.image import animation  # noqa: F401
+from pyglet.image.base import ImageData, ImageDataRegion  # noqa: F401
+from pyglet.image.animation import Animation, AnimationFrame  # noqa: F401, TC001
 
-if pyglet.options.backend in ("opengl", "gl2", "gles2"):
-    from pyglet.graphics.api.gl.image import GLBufferImage as BufferImage
-    from pyglet.graphics.api.gl.image import GLBufferImageMask as BufferImageMask
-    from pyglet.graphics.api.gl.image import GLBufferManager as BufferManager
-    from pyglet.graphics.api.gl.image import GLDepthBufferImage as DepthBufferImage
-    from pyglet.graphics.api.gl.image import GLImageData as ImageData
-    from pyglet.graphics.api.gl.image import GLImageDataRegion as ImageDataRegion
-    from pyglet.graphics.api.gl.image import GLTexture as Texture
-    from pyglet.graphics.api.gl.image import GLTexture3D as Texture3D
-    from pyglet.graphics.api.gl.image import GLTextureArray as TextureArray
-    from pyglet.graphics.api.gl.image import GLTextureArrayRegion as TextureArrayRegion
-    from pyglet.graphics.api.gl.image import GLTextureRegion as TextureRegion
-    from pyglet.graphics.api.gl.image import get_buffer_manager, get_max_texture_size
-elif pyglet.options.backend == "vulkan":
-    from pyglet.graphics.api.vulkan.image.image_temp import VulkanImageData as ImageData
-    from pyglet.graphics.api.vulkan.image.image_temp import VulkanImageData as ImageDataRegion
-    from pyglet.graphics.api.vulkan.texture import VulkanTexture as Texture
-    from pyglet.graphics.api.vulkan.texture import VulkanTextureRegion as TextureRegion
-    from pyglet.graphics.api.vulkan.texture import get_max_texture_size
-
-from pyglet.image.animation import Animation, AnimationFrame
-
-from pyglet.image.base import ImagePattern, ImageGrid, _color_as_bytes
+from pyglet.image.base import ImagePattern, ImageGrid, _color_as_bytes  # noqa: F401
 from pyglet.image.codecs import ImageDecoder
 from pyglet.image.codecs import add_default_codecs as _add_default_codecs
 from pyglet.image.codecs import registry as _codec_registry
